Add basic entry (#3)
This PR introduces the idea of an entry to the Daylio clone. It adds an entry model, migration and factory and relates it to a user.

A basic form has been added to create an entry and the index entry view shows all entries each displayed on a card. Currently, only a mood is tracked, with activities coming soon.

Some of the Tailwind styles are not working correctly, notably the background and font color for the card header when using the get_color function. I will return to this issue after activities have been introduced, as more styling will be needed for the activity card.